add a classifyIdentifier() method to Action, implement it in MinimalAction
Parsing expressions becomes easier if it's known if an identifier refers
to a type name or not. Add enough callbacks to Action so that
MinimalAction can track which identifier refers to a type (with scoping
support). Redeclarations just override the old type unconditionally,
MinimalAction won't emit redeclaration errors -- that's a job for Sema.

MinimalAction uses the usual lexer hack to annotate identifiers with
their type.

Also register all the builtin types, constants, and functions in
MinimalAction.