accept package-qualifiers
mypackage.ident could be either a type, a function, or a const/var. If
it's a type, it can be followed by '{' to form a type literal.

Knowing what it is requires enough module support to load a compiled
module and look up types by name. Since that's not available, just
always accept type literals after a package identifier and let Sema
error about this in the future.

With this, some complete real-world code snippets can be parsed!