Start working on typelit{} ambiguity in if/for/switch conditions.
No functionality change. This mostly adds a bunch of FIXMEs on how this
should work when done. There'll be a bool that flags if identifier{
should start a composite lit, and a RAII object that sets this flag to
true while in an if/for/switch condition and to false while parsing
something delimited by (), [], or {}.