correctly parse composite typename literals after if/for/switch
The go spec demands that composite typename literals between if/for/switch
and their opening '{' only appear in parentheses.  Use a RAII object to
set CompositeTypeNameLitNeedsParens to true for that range, and in that
range set the bool temporary to true in delimited expressions by
teaching BalancedDelimitedTracker about it.

The diagnostics emitted by this could be a lot better and () should be
suggested when needed, but this gets the basics right.