correctly parse parameters starting with an identifier
If the identifier started a type list not consisting solely of
identifiers, e.g. (int, somepackage.Type), this wouldn't work
previously.

Handle this by always going down ParseTypeListTail() and letting that
function return if it saw nothing but identifiers.

With this change, all files in hack/distfield parse without diagnostics.