rework how SimpleStmts are parsed, fix various bugs in there
ParseSimpleStmtTail() used to have special-case code for SimpleStmts
starting with identifier lists.  This code was duplicated in
ParseSimpleStmtTailAfterExpression() and didn't handle expression lists
that started with a simple identifier.  Now, this always parses the
start of a SimpleStmt() as expression list, and ParseExpressionList()
returns if nothing but identifiers were seen (just like it's done for
type lists).  This lets the parser accept code like `a, b.c = 2, 3`.

Also let ParseStatement() call ParseSimpleStmt() instead of
ParseExpression() for non-identifier expression prefixes, so that now
code like `(a), b = 1, 2` is accepted.

With this, the parser can parse go/src/pkg/go/parser.go!