allow package names that aren't followed by '.'
This is needed to parse `packagename := 4`, which introduces a shadowing
declaration.

ParsePrimaryExprTail() should really stop doing identifier
classification and ParseCallSuffix should instead allow a type as first
argument always.  Sema can then figure out if the call was really a
call, a conversion, or a builtin call and diag accordingly.