switch from SourceMgr::PrintMessage() to TextDiagnostic for diags
llvm::SourceMgr doesn't support fixits, and the utf8 support looks worse too.
If llvm::SourceMgr grows these capabilities in the future, this can be
switched back and the following files can be deleted:
  TextDiagnostic.h/cpp
  DiagnosticRenderer.h/cpp
  ConvertUTF.c/h and ConvertUTFWrapper.cpp
Lexer::MeasureTokenLength() and the lexer's raw mode support could
possibly be removed in that case too.

With this, fixits work!