use ExpectAndConsumeSemi() in more places
Fixes a few FIXMEs.

One side effect is that for "import (<eof>", the "expected ')'" diag
points at the next line but the "expected ';'" diag now points at the
end of the first line.  That's a bit unfortunate, but it's a bit of an
edge case and moving the ';' diag is actually a progression. The ')'
diag just needs to progress in the same way eventually.