introduce TypedIdentList, use this to build AST for ConstSpecs
Also add a few notes about more memory-efficient ways to TODO.txt. For
now, get something working first.

TypedIdentList is a list of idents with a type (`a, b, c int`) and for
now a base class for VarSpecDecl and the new ConstSpecDecl.

Use this to build AST for const, and use that to emit diags for consts
from ActOnTypeName() (fixes a FIXME).