implement field lookup for structs
Only works for "direct" types for now, e.g. `var foo struct {...}`, not
if the struct is a named type.

Makes FieldDecl a transparent decl context and pushes it on the decl
stack. This also required letting the parser create a parse scope for
every field decl.

Doesn't handle embedded fields yet.