don't crash on field access in invalid base expressions
The rule is that the Parser must never call Action methods with an
isInvalid() argument. (TODO: audit the parser code for this, this is
currently not true everywhere.)

Empty but valid base expressions (caused by not yet implemented
Action methods that return ExprEmpty() in their default implementation)
still cause crashes, but these will go away when more of Sema is
implemented.