start implementing assignability checking
See http://tip.golang.org/ref/spec#Assignability for the rules.

Types are considered identical if they point to the same object. It is
the responsibility of ASTContext::getTypeDeclType() to convert decls to
identical types according to the rules described in
http://tip.golang.org/ref/spec#Type_identity . Also update that function
to return identical Type objects for NameTypes coming from the same
TypeSpecDecl. (The function will be fixed for more Decl objects in
follow-up commits.)

Assignability checking is only implemented for named types for now,
further checks will be added in follow-up commits.